Changed everything to doubles, moved rendering to render function

+++ Heros/Actor.cpp
@@ -5,14 +5,14 @@ Actor::Actor() : GameObject() {}
 
 Actor::Actor(Vector2 pos, Vector2 size) : GameObject(pos, size) {}
 
-void Actor::Falling(Vector2 grav, float delta){
+void Actor::Falling(Vector2 grav, double delta){
 	velocity_ += (grav * delta);
 }
 
 
 void Actor::EnteredCollision(GameObject *object, Vector2 &overlapped)
 {
-	if (overlapped.y < (float)0)
+	if (overlapped.y < 0)
 		onGround = true;
 
 	//position_ += overlapped;
