+++ Heros/Player.h
@@ -5,7 +5,7 @@
 class Player : public Actor {
 public:
     Player(Vector2 position, Vector2 size);
-    void Update(float, Keys);
+    void Update(double, Keys);
 	virtual void EnteredCollision(GameObject *, Vector2 &);
     void Render(Viewport &);
 };
